control-plane/ap-controller/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
redis==5.0.8
PyYAML==6.0.2
